# ezitech-lms-server/openapi.yaml
openapi: 3.0.3
info:
  title: Ezitech LMS API
  version: "1.0"
  description: |
    Enterprise Learning Management System API. All routes are mounted
    under `/api/{version}` (see `servers` below) - the version segment
    matches `API_VERSION` in the server's environment config (`v1` by
    default).

    ## Authentication
    Most endpoints require a Bearer access token obtained from
    `POST /auth/login` (or `/auth/refresh`). A handful of read-only
    endpoints - the public course catalog, certificate verification,
    the subscription plan list, course review listings - accept an
    OPTIONAL token: an anonymous caller gets a public view, an
    authenticated one gets a personalized/entitled view of the same
    resource. Each such path says so explicitly below.

    ## Response envelope
    Every endpoint responds with the same envelope shape
    (`ApiSuccess`/`ApiError` below) regardless of route - this API does
    not return bare arrays or bare objects at the top level anywhere.

    ## What this spec does NOT attempt
    This was hand-authored against the actual route/controller/
    validation-schema source rather than generated from code
    annotations, so it has not been validated by a schema linter or
    exercised against the live API. Request/response schemas favor
    accuracy on required fields, enums, and auth/error shapes over
    exhaustively modeling every optional field on every lesson or
...

  contact:
    name: Ezitech Engineering
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /courses/{courseId}:
    get:
      tags: [Courses]
      summary: Get a course
      description: 404s (not 403) for a draft/archived course unless the caller owns it or is admin/course_manager - existence isn't disclosed to anyone else.
      security: []
      parameters: [{ $ref: '#/components/parameters/courseId' }]
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/Course' } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Courses]
      summary: Update a course (owner/admin/course_manager only)
      parameters: [{ $ref: '#/components/parameters/courseId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Same shape as create, all fields optional. }
      responses:
        '200': { description: Updated. }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Courses]
      summary: Delete a course (owner/admin/course_manager only)
      description: Cascades to sections, lessons, assessments, submissions, enrollments, progress, live sessions, attendance, SCORM data, discussion threads/comments and reviews, and removes the course from learning paths and other courses' prerequisites. Uploaded files are deleted best-effort. Previously-issued certificates are preserved with the course title stored on them. Refused with 409 when any student paid for the course - archive it instead.
      parameters: [{ $ref: '#/components/parameters/courseId' }]
      responses:
        '200': { description: Deleted. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: The course has paying students. }
...
  /subscriptions/plans:
    get:
      tags: [Subscriptions]
      summary: List subscription plans (public catalog)
      security: []
      responses: { '200': { description: OK } }
    post:
      tags: [Subscriptions]
      summary: Create a subscription plan (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, durationDays]
              properties:
                name: { type: string, maxLength: 200 }
                description: { type: string, maxLength: 2000 }
                price: { type: number, minimum: 0 }
                durationDays: { type: integer, minimum: 1 }
                coursesIncluded:
                  oneOf:
                    - { type: string, enum: [all] }
                    - { type: array, items: { $ref: '#/components/schemas/ObjectId' } }
                isActive: { type: boolean, default: true }
      responses: { '201': { description: Created. }, '403': { $ref: '#/components/responses/Forbidden' } }
  /subscriptions/plans/{planId}:
    patch:
      tags: [Subscriptions]
      summary: Update a subscription plan (admin only)
      parameters:
        - in: path
          name: planId
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
      requestBody:
        content:
          application/json:
            schema: { type: object, description: Same shape as create, all fields optional. }
      responses: { '200': { description: Updated. }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /subscriptions/subscribe:
    post:
      tags: [Subscriptions]
      summary: Subscribe to a plan
      description: Activates the plan immediately; no payment gateway is involved. Answers 402 only when the server runs with PAYMENTS_ENABLED=false.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties: { planId: { $ref: '#/components/schemas/ObjectId' } }
      responses: { '201': { description: Subscribed. } }
  /subscriptions/{subscriptionId}/cancel:
    patch:
      tags: [Subscriptions]
      summary: Cancel a subscription
      parameters:
        - in: path
          name: subscriptionId
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
      responses: { '200': { description: Cancelled. } }

  # ============================== LIVE CLASSES ==============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Obtained from `POST /auth/login` or `/auth/refresh`. Send as `Authorization: Bearer <token>`.'
  parameters:
    courseId: { name: courseId, in: path, required: true, schema: { $ref: '#/components/schemas/ObjectId' } }
...
  responses:
...
    Forbidden:
      description: Authenticated, but not permitted to perform this action.
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: Resource not found (also returned instead of 403 where existence itself shouldn't be disclosed - see individual endpoint notes).
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  schemas:
    ObjectId:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
      example: 64f1a2b3c4d5e6f7a8b9c0d1
...
    ApiSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, nullable: true }
        data: {}
        pagination:
          type: object
          nullable: true
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            pages: { type: integer }
    ApiError:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        details:
          type: array
          items: { type: string }
          description: Present on validation errors (400) - one entry per failing field.
...
    Course:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        instructor: { $ref: '#/components/schemas/ObjectId' }
        category: { $ref: '#/components/schemas/ObjectId' }
        status: { type: string, enum: [draft, published, archived] }
        price: { type: number }
        tags: { type: array, items: { type: string } }
        prerequisites: { type: array, items: { $ref: '#/components/schemas/ObjectId' } }
        contentVersion: { type: integer }
        createdAt: { type: string, format: date-time }
...